Refactor them into respective files
